requirements.txt
aiohttp==1.0.5
arrow==0.10.0
async-timeout==2.0.0
certifi==2017.11.5
chardet==3.0.4
discord.py==0.16.12
gunicorn==19.7.1
idna==2.6
multidict==3.3.2
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
websockets==3.4
